create_trackplot creates a horizontal plot with rows representing separate tracks

create_trackplot creates a horizontal plot with rows representing separate tracks

Usage

create_trackplot(
  track_data,
  row_id,
  metric_id,
  xlimits = c(min(track_data$start), max(track_data$end)),
  colour.scheme = c("slateblue4", "violetred3"),
  total.colours = 99,
  colour.centering.value = 0,
  colour.alpha = 1,
  fill.colour = "white",
  at = NULL,
  print.colour.key = TRUE,
  colourkey.labels.at = NULL,
  colourkey.labels = NULL,
  filename = NULL,
  main = list(label = ""),
  main.just = "center",
  main.x = 0.5,
  main.y = 0.5,
  main.cex = 3,
  yaxis.lab = NULL,
  xaxis.lab = NULL,
  xaxis.lab.top = NULL,
  xaxis.cex = 1.5,
  xaxis.top.cex = NULL,
  yaxis.cex = 1.5,
  xlab.cex = 2,
  ylab.cex = 2,
  xlab.top.label = NULL,
  xlab.top.cex = 2,
  xlab.top.col = "black",
  xlab.top.just = "center",
  xlab.top.x = 0.5,
  xlab.top.y = 0,
  xat = TRUE,
  xat.top = NULL,
  yat = TRUE,
  xaxis.tck = NULL,
  xaxis.top.tck = NULL,
  yaxis.tck = NULL,
  xaxis.col = "black",
  yaxis.col = "black",
  col.pos = NULL,
  row.pos = NULL,
  cell.text = "",
  text.fontface = 1,
  text.cex = 1,
  text.col = "black",
  text.position = NULL,
  text.offset = 0,
  text.use.grid.coordinates = TRUE,
  colourkey.cex = 3.6,
  xaxis.rot = 90,
  xaxis.rot.top = 90,
  yaxis.rot = 0,
  xlab.label = "",
  ylab.label = "",
  xlab.col = "black",
  ylab.col = "black",
  axes.lwd = 2,
  gridline.order = "h",
  grid.row = FALSE,
  grid.col = FALSE,
  force.grid.row = FALSE,
  force.grid.col = FALSE,
  grid.limit = 50,
  row.lines = NULL,
  col.lines = NULL,
  top.padding = 0.1,
  bottom.padding = 0.5,
  right.padding = 0.5,
  left.padding = 0.5,
  x.alternating = 1,
  shrink = 1,
  row.colour = "black",
  col.colour = "black",
  row.lwd = 1,
  col.lwd = 1,
  grid.colour = NULL,
  grid.lwd = NULL,
  width = 6,
  height = 6,
  size.units = "in",
  resolution = 1600,
  enable.warnings = FALSE,
  xaxis.fontface = "bold",
  yaxis.fontface = "bold",
  input.colours = FALSE,
  axis.xlab.padding = 0.1,
  ...
)

Arguments

track_data

A dataframe detailing segments

start

the start position of the segment, based on histogram vector index

end

the end position of the segment, based on histogram vector index

`metric_id`

values; factor or continuous data for plotting

`row_id`

character or factor; identifier for the row

colour

colour for the row

row_id

character, column name used to indicate the track row

metric_id

character, column name used to scale colour

xlimits

numeric vector of length 2 indicating the start and end indices of the track respectively

colour.scheme

Heatmap colouring. Accepts old-style themes, or a vector of either two or three colours that are gradiated to create the final palette.

total.colours

Total number of colours to plot

colour.centering.value

What should be the center of the colour-map

colour.alpha

Bias to be added to colour selection (uses x^colour.alpha in mapping). Set to “automatic” for auto-adjustment.

fill.colour

The background fill (only exposed where missing values are present

at

A vector specifying the breakpoints along the range of x; each interval specified by these breakpoints are assigned to a colour from the palette. Defaults to NULL, which corresponds to the range of x being divided into total.colours equally spaced intervals. If x has values outside of the range specified by “at” those values are shown with the colours corresponding to the extreme ends of the colour spectrum and a warning is given.

print.colour.key

Should the colour key be printed at all?

colourkey.labels.at

A vector specifying the tick-positions on the colourkey

colourkey.labels

A vector specifying tick-labels of the colourkey

filename

Filename for tiff output, or if NULL returns the trellis object itself

main

The main title for the plot (space is reclaimed if NULL)

main.just

The justification of the main title for the plot, default is centered

main.x

The x location of the main title, deault is 0.5

main.y

The y location of the main title, default is 0.5

main.cex

Size of text for main plot title, defaults to 2.5

yaxis.lab

A vector of col labels, NA = use existing colnames, NULL = none

xaxis.lab

A vector of row labels, NA = use existing rownames, NULL = none

xaxis.lab.top

The label for the top x-axis. Required only if you want to print a top *and* bottom xaxis, otherwise use x.alternating = 2 for top axis only. Defaults to NULL

xaxis.cex

Size of x-axis label text - defaults to values found in a look-up table

xaxis.top.cex

Size of top x axis label text

yaxis.cex

Size of y-axis label text - defaults to values found in a look-up table

xlab.cex

Size of x-axis label, defaults to 2

ylab.cex

Size of y-axis label, defaults to 2

xlab.top.label

The label for the top x-axis

xlab.top.cex

Size of top x-axis label

xlab.top.col

Colour of the top x-axis label

xlab.top.just

Justification of the top x-axis label, defaults to centered

xlab.top.x

The x location of the top x-axis label

xlab.top.y

The y location of the top y-axis label

xat

Vector listing where the x-axis labels should be drawn, defaults to automatic

xat.top

Vector listing where the x-axis labels should be drawn on the top of the plot. Required only when you want bottom and top axis, otherwise use x.alternating = 2, to get top axis only. Defaults to NULL

yat

Vector listing where the y-axis labels should be drawn, defaults to automatic

xaxis.tck

Size of x-axis tick marks. Defaults to NULL for intelligent choice based on covariate size.

xaxis.top.tck

Size of top x-axis tick marks. Defaults to NULL for intelligent choice based on covariate size.

yaxis.tck

Size of y-axis tick marks. Defaults to NULL for intelligent choice based on covariate size.

xaxis.col

Colour of the x-axis tick labels, defaults to “black”

yaxis.col

Colour of the y-axis tick labels, defaults to “black”

col.pos

Vector of column positions for adding text to cell, defaults to NULL

row.pos

Vector of row positions for adding text to cell, defaults to NULL

cell.text

Text to add to cell, defaults to an empty string

text.fontface

1 = Plain, 2 = Bold, 3 = Italic, default is 1

text.cex

Text size, default is 1

text.col

Text colour, default is black.

text.position

The position of the text, defaults to center.

text.offset

The offset of the position, defaults to 0.

text.use.grid.coordinates

Indetifier if grid coordinates or npc coordinates should be used

colourkey.cex

Size of colourkey label text

xaxis.rot

Rotation of x-axis tick labels; defaults to 90

xaxis.rot.top

Rotation of the top x-axis tick labels; defaults to 90

yaxis.rot

Rotation of y-axis tick labels; defaults to 0

xlab.label

The label for the x-axis

ylab.label

The label for the y-axis

xlab.col

Colour of the x-axis label, defaults to “black”

ylab.col

Colour of the y-axis label, defaults to “black”

axes.lwd

Width of heatmap border. Note it also changes the colourkey border and ticks

gridline.order

Character specifying order in which to draw interior grid-lines ('h' or 'v'). Defaults to 'h' for horizontal first.

grid.row

Allow turning off of the interior grid-lines. Default FALSE

grid.col

Allow turning off of the interior grid-lines. Default FALSE

force.grid.row

Overrides default behaviour of turning off grid lines when number of rows exceed grid.limit. Defaults to FALSE

force.grid.col

Overrides default behaviour of turning off grid lines when number of columns exceed grid.limit. Defaults to FALSE

grid.limit

Limit set for when to turn off column and row lines if data size exceeds it. Defaults to 50

row.lines

Vector specifying location of lines, default is seq(1, ncol(x), 1) + 0.5. Note: Add 0.5 to customized vector

col.lines

Vector specifying location of lines, default is seq(1, nrow(x), 1) + 0.5. Note: Add 0.5 to customized vector

top.padding

A number specifying the distance to the top margin, defaults to 0.1

bottom.padding

A number specifying the distance to the bottom margin, defaults to 0.5

right.padding

A number specifying the distance to the right margin, defaults to 0.5

left.padding

A number specifying the distance to the left margin, defaults to 0.5

x.alternating

A value specifying the position of the col names, defaults to 1. 1 means below the graph, 2 means above the graph. Use 3 to get tick marks below and above graph, but still need to specify xat.top and xaxis.lab.top to get values there

shrink

Allows rectangles to be scaled, defaults to 1

row.colour

Interior grid-line colour, defaults to “black”. Can be a vector

col.colour

Interior grid-line colour, defaults to “black”. Can be a vector

row.lwd

Interior grid-line width, defaults to 1. Setting to zero is equivalent to grid.row = FALSE and grid.col = FALSE. Can be a vector.

col.lwd

Interior grid-line width, defaults to 1. Setting to zero is equivalent to grid.row = FALSE and grid.col = FALSE. Can be a vector.

grid.colour

Interior grid-line colour, defaults to “black”. Can be a vector. Applies to both rows and columns. DEPRECATED

grid.lwd

Interior grid-line width, defaults to 1. Setting to zero is equivalent to grid.row = FALSE and grid.col = FALSE. Applies to both rows and columns. DEPRECATED

width

Figure width in size.units

height

Figure height in size.units

size.units

Units of size for the figure

resolution

Figure resolution in dpi

enable.warnings

Print warnings if set to TRUE, defaults to FALSE

xaxis.fontface

Fontface for the x-axis scales

yaxis.fontface

Fontface for the y-axis scales

input.colours

boolean expressing whether or not the matrix was specified using colours or integer values. Defaults to FALSE

axis.xlab.padding

Padding between axis of plot and x label

...

Additional arguments to be passed to BoutrosLab.plottin.general::create.heatmap

Value

Track plot, a Trellis object. For further details, see the 'Lattice' R package.

Examples

if (FALSE) {
x = Histogram(c(0, 0, 1, 2, 3, 2, 1, 2, 3, 4, 5, 3, 1, 0))
results = segment_and_fit(x, eps = 0.005)
results_table = summarize_results(results)
create_trackplot(
track_data = results_table,
row_id = "peak_id",
metric_id = "value"
)
}